Extract `PluginSites` component so it can be reused on mobile view
