+++ client/dashboard/plugins/manage/components/plugin-sites.scss
